requirements.txt
boto3==1.17.112
botocore==1.20.112
futures==3.3.0
jmespath==0.10.0
python-dateutil==2.8.2
s3transfer==0.4.2
six==1.16.0
urllib3==1.26.6
atomicwrites==1.4.0
pathlib_mate==1.0.1
pycodestyle==2.6.0
autopep8==1.4.4
scandir==1.10.0
toml==0.10.2
attrs==20.3.0
sqlitedict==1.7.0
diskcache==4.1.0
certifi==2019.11.28
idna==2.10
chardet==4.0.0
requests==2.25.1
typing==3.7.4.3
whoosh==2.7.4
PyYAML==5.4.1
